Price is decimal, not integer

+++ db/migrate/20230607132451_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
   def change
     create_table :orders do |t|
       t.string :kind, null: false
-      t.integer :price, null: false
+      t.decimal :price, null: false
       t.string :customer, null: false
 
       t.timestamps
